services/image-server/requirements.txt
Flask==1.1.*
mysql-connector-python==8.0.*
flask-cors==3.0.*
requests==2.24.*

